Fix problem to restore back previous alignments on tabbed paged header text block.
When showing images in tabbed paged headers we store old alignment of text block
to be able to restore when hiding images (getting back original alignment of text block).
If sizes where changed for images while showed, the old values where lost.

Fix makes sure only the old values are stored, if current alignment is not the
default (center) when displaying images in tabbed paged header.