Adding LineHeight to Label and Span
Fixes #1734 by adding LineHeight property to Label and Span.

Setting LineHeight on Span doesn't work on UWP ince it's not supported
on the Run class corresponding to Span in Xamarin Forms.

On Android the separate LineHeight on span only has effect if
the text has linebreaks within the string. Then it's possible
to set different line heights for individual lines using Span.
If the lines are automatically split using WordWrap for instance,
it seems like Android will use the LineHeight set on the last Span.