Implemented review feedback.
* Animation problems related to Android API level 19 and earlier
due to a couple of bugs and changed behavior in AnimationDrawable class.
Implemented a couple of fixes and workaround related to these bugs, fixing
the animation issues on devices before Android Lollipop.

* Moved a couple of methods implemented in FormsAnimationDrawable into an
interface that is used to hide implementation details. Enables possibility
to factor different implementations of animation support going forward.

* Fixed review feedback and comments.