[Flex] Make sure layout is invalidated on time
When a BP changes, the BP value changes, the propertyChanged event is
fired, then the propertyChanged delegate is executed.

As the invalidation was done in the event, and setting the flexitem
property was done in the delegate, the invalidation, and relayout, was
executed with an out-of-date flexitem.

With this change, we invalidate the child size, and that'll bubble up to
invalidate the layout, as part of the property changed delegate

- fixes #2821