Add basic application key shortcuts
A Xam.Forms standard NSApplication override is necessary to allow basic keyboard shortcuts while not using any application menus. The FormsApplication extends NSApplication while overriding the SendEvent method to be able to intercept key events to enable basic keyboard command functionality for text fields and views. The FormsApplication is used as the Principal class attribute in a project Info.plist. Until the option is added as a VS for mac item, the Info.plist file must be opened using XCode to alter the value