[CSS] allow mapping css prop to multiple BPs
To allow mapping a single CSS property to multiple controls, we so far
relied on extracting the BindableProperties to static helper classes,
and implementing an interface for accessors, and eventually change
handlers.

That extraction works well, and is actually a good pattern, as long as:
 - the extracted BPs share the same PropertyName,
 - default value, and
 - return type, obiously

As the CSS `color` property has to map to both `TextColor` and `Color`
BPs, the property extraction wasn't possible.

This change adds the capability to map a single CSS property to
multiples BPS, when extraction isn't possible. Whenever the extraction
to an interface is possible, that behavior is strongly encouraged.