[C] Expose Application.LogWarningsListener
The fix proposed in #2911 didn't worked with Release build of XF, aka
the majority of the builds used.
This fix allows the developer to set any logger he might like, but the
plan is to add a default one in the template, that uses Debug.WriteLine,
so it will Log if the user application is build with Debug, and not if
XF binaries are.

- fixes #1517
- fixes #3873