[build] add a stamp file to shamehat
In addition to #4579 that check, in the tool, that the extra
method isn't present, this protects the task itself from being ran
twice.

Both fixes are required for extra performance and safety.

This also removes the last reference to the original name of the tool.
But I'll still remember it fondly, and use it in internal conversations.