Add Device.IsVirtual property
Sometimes (especially for tooling and/or telemetry purposes) it's
useful to know if your code is running on a physical device or an
emulator/simulator. Xamarin.Essentials provides this information
via the DeviceInfo class, but it requires its own separate initialization
in addition to what XF already does.

For a library targeting XF, depending on Xamarin.Essentials to provide
this information would mean taking a dependency on its package.
Consumers of the library would need to now know to initialize Essentials
properly or otherwise the library will fail. Since it would be a
transitive dependency, it wouldn't be easy to figure out what's
required to make it work.

So in order to make this information available in the easiest
possible way, a new `IsVirtual` property is added to the `Device`
class, which is set for Android and iOS in the same way Essentials does,
but using a simple boolean instead to reduce the added API surface.