Added test to cover setting a property that throws
Currently we (meaning the Forms Previewer) get notified if an exception is thrown - by Forms code that processes setting property values from the XAML file - because a property is in some way invalid - that is, the XAML is invalid - the property doesn't exist, or it's the wrong type or whatever.

However, if the XAML is valid and Forms actually sets the property, but an exception occurs in the code that handles setting the property, we aren't notified and the XAML does not get inflated. If possible, we'd like to be notified in this scenario, and the exception be otherwise ignored.