Adds a test to verify root node FallbackTypeInfos
If the root tag is something like <local:Foo>, Forms doesn't generate any FallbackTypeInfos, so the previewer has no way to process the type.

This seems to be because of these lines in XamlParser.GetTypeReference()...

				if (defaultAssemblyName == null)
					return null;

When determining the type of the root tag, defaultAssemblyName is null because Context.RootElement (referenced in CreateValuesVisitor.Visit to get the root assembly) has not yet been set.