Fix [Android] WebView Navigating Cancel property not working with custom scheme #4891
Updates Xmarin.Forms.Platform.Android.FormsWebViewClient

This was broken with release 3.3.0.

Notes:
Depending on the API version, Android's StopLoading does not work as expected/consistently which is used to cancel the load in OnPageStarted. This is why 3.3.0 created this 'custom scheme' problem.

In addition, Android's ShouldOverrideUrlLoading is NOT executed when the WebView.Source is explicitly set. This is why 3.3.0 fixed the problem of Navigating not firing when the Source is set.

### Issues Resolved ###
fixes #4891

### API Changes ###
None

### Platforms Affected ###
Android

### Behavioral/Visual Changes ###
None

### Before/After Screenshots ###
Not applicable

### Testing Procedure ###
see Issue4891.cs