Ignore FindByName exception when running in previewer
When running in the previewer, some types are replaced. If a replaced type as an x:Name attribute, this will result in an InvalidCastException when FindByName tries to cast it to T. So return default(T) in this scenario (this may just result in an exception later trying to use the result of this call, but it is at least less likely to be in the constructor).