Resolve fallback type before trying to make constructed generic type
The FallbackTypeResolver hander in the Forms previewer does not have access to the x:TypeArguments attribute, so if it constructs a proxy generic type from metadata it can't construct it. So the previewer has to create an unconstructed generic type, and Forms will need to construct it.

So we need to move the code that constructs the generic type after the call to FallbackTypeResolver.

(cherry picked from commit 5acfea4f457453856a8905002f4d35c8e9aad106)