Use correct reference frame for gesture origin recognition
UIGestureRecognizer.LocationInView(null) defaults to the global window view as reference frame, whose coordinate system is in portrait mode, always.
Thus we need to query the root element whose coordinate system is adapted to the current rotation (i.e. portait / landscape).
Which, in this context, is UIApplication.SharedApplication.KeyWindow.RootViewController.View.

This fixes in particular installed TagGestureHandlers on Span elements  (of a FormattedString of a Label for example) in landscape mode.