[Android] fix animations not animating
fixes #7255
fixes #7291
fixes #7278

I found that if the Animation duration scale setting has never been changed in the developer settings, the scale we get will allways be 0. So instead of defaulting to 0 we default to 1. This means if i havent touched the setting it will behave as it did pre 4.2.

Also if somehow the context is null, we should default to enabling animations since this is far more "normal" than disabling animations.