[C] Fix TemplateBinding for non-logicalchildren
Spans, e.g., are no LogicalChildren of FormattedString, which is not a
LogicalChildren of Label. The newly introduced way of propagating the
TemplatedParent was top-down, using LogicalChildren, was not assigning
the correct TemplatedParent on Spans.

This reverts to a bottom-up lookup, hooking on ParentSet, like it used to
be.

As a side effect, the TemplateBindings are no longer updated on reparenting,
but this was the original behavior, and I can't think of a case where this
would be needed.

The regression to `{TemplateBinding}` was introduced by #4375.

- fixes #7494