[Core] remove array covariant cast for UWP
Fixes: https://github.com/xamarin/Xamarin.Forms/issues/9134

A reported failure of:

    System.InvalidCastException: 'Unable to cast object of type 'System.Attribute[]' to type 'Xamarin.Forms.HandlerAttribute[]'.'

was happening on UWP, @PureWeen also could repro with a new app.

What is weird about this:

* It appears to work with .NET Framework & Mono
* It appears to work for me locally if I just launch the Control
  Gallery on UWP.

I suspect something is happening with how Xamarin.Forms is built
(using 2017 and older UWP SDKs), but not really sure?

I think we should just fix this quickly and I will revisit later.