Fix #9711 Failed to marshal HeaderWrapperView
Crash appears to be occurring due to the premature disposal of a `HeaderWrapperView` instance.

Crash occurs in `VisualElementRenderer::UpdateParentPageAccessibilityElements()`. As it walks up the superview chain from the `UILabel` that was tapped on in the test, it arrives at a superview whose `NSObject` cannot be retrieved. That superview happens to be a `HeaderWrapperView` instance. In `Runtime.cs:1249`, the call to `TryGetNSObject()` to retrieve this `HeaderWrapperView` returns `null`.  In the debugger, I manually called `TryGetNSObject()` with `true` for the `evenInFinalizerQueue` parameter, which returned the expected instance -- confirming that the instance has been unexpectedly garbage collected.

The resulting exception claims that it was not "possible to create a new managed instance (because the type 'Xamarin.Forms.Platform.iOS.HeaderWrapperView' does not have a constructor that takes one IntPtr argument)".  In my first attempted workaround, I added the missing constructors.  This got me past the initial problem, only to reveal another -- a call to `ShouldReceiveTouch()` with a `HeaderWrapperView` instance passed as the `UITouch` parameter.  That looked to me like a dangling pointer problem, so I decided to look elsewhere.

I'm very new to Xamarin/.NET, so someone with more context might be able to discover why this `HeaderWrapperView` instance was garbage collected. But I do know iOS, and when I looked into where and how `HeaderWrapperView` is used I saw some suspect code.  Typically, when implementing section headers with custom views in a `UITableView`, you want to use the same kind of reuse that you do for table view cells.  That is, you should use `UITableView.dequeueReusableHeaderFooterView()` to retrieve the section header view.  However, the current implementation creates a new `HeaderWrapperView` on each call to `ListViewDataSource.GetViewForHeader()`. In the test case, this happens each time the state changes in response to the tap on a section header.

To work around the problem, I used `UITableView.dequeueReusableHeaderFooterView()` to attempt reuse of `HeaderFooterView` so a new instance isn't created each time.  This appears to be sufficient to avoid the original issue.

**Additional considerations**

This header reuse only applies to `HeaderWrapperView`, and not to its subviews.  It would be better for scrolling performance if all subviews of `HeaderWrapperView` could be reused. I attempted to implement that (See commented out line at `ListViewRenderer.cs:1122`), but was unable to get the subviews to refresh properly. Again someone with more Xamarin experience and context might be able to make this work.