Update LabelRenderer.cs
The height of the Label seems to be off by 1 pixel, most likely due to a rounding error further down in the XF framework.

As such, when `layoutManager.BoundingRectForGlyphRange` is called for anything on the last line, it will return 0, because the glyph is outside the bounds. By adding the 1 extra pixel, it brings the glyph inside the bounds and correctly returns the CharacterBounds.

If it fails to determine a region, then any GestureRecognizer will fail on the Span, which is what happens if the GestureRecognizer is for a span on the last line of a Label.