Update LabelExtensions.cs
This update only updates the region to get the range, rather than modify the actual Span size.

Second, the initial size comes in a x = -1, y = -1 and width = 1, and height = 1. So changing the return to 1, instead of 0, allows an exit quicker, rather than calculating regions for 1px x 1px container sizes. You aren't clicking that anyway.