Fixed ScrollY when ItemsSource reloaded or row height changed
Fixes #8305 and #9370 and another bug that has not been reported - changing height of the first row also creates incorrect ScrollY value.
Not 100% sure if there can't be any other situation when the scroll fails to be tracked, for which setting the  `_contentOffset` to `0` might not be correct, but it fixes the "reloaded ItemsSource" issue, which seems to be the most common. When we were not able to track the offset at all, it's a bug anyway, so setting the offset to 0 seems good because it solves this issue at least.