Add an example of using mdoc(5) XML with <include/>
Suppose you want to have C# `///` docs as the "source of truth", but
you don't want to work quite so hard to *import* those docs into
C# `///` format.

This is where the [`<include/>`][0] documentation comment comes in:
since **mdoc**(5) documentation is XML, you can *copy* the `.xml`
files into this repo, and then use an XPath expression to extract
documentation from the **mdoc**(5) documentation:

	<!-- file: On.xml -->
	<Type Name="On" FullName="Xamarin.Forms.On">
	  <Docs>
	    <summary>Class that is used within <c>OnPlatform</c> tags in XAML when specifying values on platforms.</summary>
	  </Docs>
	  <Members>
	    <Member MemberName=".ctor">
	      <Docs>
	        <summary>Creates a new <see cref="T:Xamarin.Forms.On" /> with default values.</summary>
	…

	// File: On.cs

	/// <include file="On.xml" path="/Type/Docs" />
	public partial class On {
	    /// <include file="On.xml" path="/Type/Members/Member[@MemberName='.ctor']/Docs" />
	    public On()
	    {
	    }
	}

This results in a somewhat "conceptually weird" workflow:

 1. Documentation is in `On.xml`, in **mdoc**(5) format.

 2. `msbuild Xamarin.Forms.Core.csproj` generates a
    `Xamarin.Forms.Core.xml` file which contains docs from `On.xml`

 3. `make Xamarin.Forms.Core.Docs` /
    `mdoc update --import Xamarin.Forms.Core.xml …` generates *new*
    **mdoc**(5) output XML for integration w/ Xamarin.Forms-api-docs,
    based on the contents of `Xamarin.Forms.Core.xml`.

While this feels "weird" -- we start with **mdoc**(5) XML, and wind up
with **mdoc**(5) XML! -- it's useful as *transitionary* infrastructure,
making it easier to begin using C# `///` documentation comments.

Additionally, using C# `///` comments as an "intermediary" also means
that (1) doesn't *need* to be *strictly* **mdoc**(5) format.
One issue with **mdoc**(5) XML files is that method overloads are
difficult to query via XPath.  For example, consider
`AbsoluteLayout.OnChildRemoved`:

	<Type Name="AbsoluteLayout" FullName="Xamarin.Forms.AbsoluteLayout">
	  <Members>
	    <Member MemberName="OnChildRemoved">
	      <MemberSignature Language="C#" Value="protected override void OnChildRemoved (Xamarin.Forms.Element child);" />
	      <Docs>…</Docs>
	    </Member>
	    <Member MemberName="OnChildRemoved">
	      <MemberSignature Language="C#" Value="protected override void OnChildRemoved (Xamarin.Forms.Element child, int oldLogicalIndex);" />
	      <Docs>…</Docs>

They both share the same `//Member/@MemberName` value, and differ in
`//Member/MemberSignature/@Value` values and `//Member/Parameters`
values.  While they *can* be distinguished in XPath, it would be much
easier for C# `///` migration purposes to *change* the
`//Member/@MemberName` value so that they're unique, making for an
easier XPath query:

	<Type Name="AbsoluteLayout" FullName="Xamarin.Forms.AbsoluteLayout">
	  <Members>
	    <Member MemberName="OnChildRemoved">
	      <MemberSignature Language="C#" Value="protected override void OnChildRemoved (Xamarin.Forms.Element child);" />
	    </Member>
	    <Member MemberName="OnChildRemoved2">
	      <MemberSignature Language="C#" Value="protected override void OnChildRemoved (Xamarin.Forms.Element child, int oldLogicalIndex);" />

Or, the XML can be radically simplified, if desired:

	<Members>
	  <OnChildRemoved>
	    <!-- contents of previous Docs/ element -->
	  </OnChildRemoved>
	  <OnChildRemoved2>
	    <!-- contents of previous Docs/ element -->
	  </OnChildRemoved2>

The point is, step (2) provides a degree of flexibility around "where"
documentation is located, flexibility that isn't present when using
only **mdoc**(5) documentation.

[0]: https://docs.microsoft.com/dotnet/csharp/language-reference/language-specification/documentation-comments#include