Added Cancel Task To WebNavigatingEventArgs
Added a new property which is a Func that returns a Task<bool>. This can be implemented in the Navigating event fired by the WebView and will allow a developer to pass the native renderer some code to execute which will evaluate whether or not to cancel the current navigation. The WebView will wait for this code to execute before continuing.