ListView: avoid that disabling RefreshAllowed cancels refresh indicator on Android, fixes #8384
RefreshAllowed is bound to ListView.RefreshCommand.CanExecute().
Often an implementation of RefreshCommand might update its CanExecute status
after execution starts. This caused ListViewRenderer to immediately
disable the SwipeRefreshLayout, thereby cancelling the refresh/activity
indicator on top of the list view.

The solution is to NOT disable it while refreshing, but waiting for the
next chance when current refresh activity/command is done.