ListView: avoid that disabling RefreshAllowed cancels refresh indicator on Android, fixes #8384 (#14816)
* Add test for issue #8384

* ListView: avoid that disabling RefreshAllowed cancels refresh
indicator on Android, fixes #8384

RefreshAllowed is bound to ListView.RefreshCommand.CanExecute().
Often an implementation of RefreshCommand might update its CanExecute status
after execution starts. This caused ListViewRenderer to immediately
disable the SwipeRefreshLayout, thereby cancelling the refresh/activity
indicator on top of the list view.

The solution is to NOT disable it while refreshing, but waiting for the
next chance when current refresh activity/command is done.

* Only enable Issue8384 constructor for XF controls app build (not for UI unit tests)

Otherwise we get 'InitializeComponent' not found error when compiling UI unit test projects.